requirements/testing.txt
pysparkling
psycopg2-binary
-e /usr/local/lib/jobs
-e /opt/spark/python